move fastformulaareas into its own directory
